requirements.txt
numpy==1.24.3
tensorflow==2.13.0
Flask==3.0.0
Flask-Uploads==0.2.1
Werkzeug==3.0.1
Pillow==10.0.1